+++ src/applications/health-care-questionnaire/questionnaire/api/index.js
@@ -1,25 +1,4 @@
-import { apiRequest } from 'platform/utilities/api';
-import environment from 'platform/utilities/environment';
+import { submit, transformForSubmit } from './submit-questionnaire';
+import { loadAppointment } from './load-appointments';
 
-const USE_MOCK_DATA = window.Cypress || environment.isLocalhost();
-
-const loadAppointment = async () => {
-  let promise;
-  if (USE_MOCK_DATA) {
-    promise = new Promise(resolve => {
-      setTimeout(() => {
-        import(/* webpackChunkName: "appointment-data" */ './appointment-data.json').then(
-          module => {
-            resolve(module.default);
-          },
-        );
-      }, 0);
-    });
-  } else {
-    const url = '/health_quest/v0/appointments/123123';
-    promise = apiRequest(`${environment.API_URL}${url}`);
-  }
-  return promise;
-};
-
-export { loadAppointment };
+export { loadAppointment, submit, transformForSubmit };
