+++ src/applications/health-care-questionnaire/questionnaire/tests/unit/reducer/reducer.loading.appointment.data.unit.spec.js
@@ -30,5 +30,6 @@ describe('health care questionnaire -- questionnaire reducer --', () => {
       state.context.appointment.attributes.vdsAppointments.length,
     ).to.be.equal(1);
     expect(state.context.appointment.attributes.vdsAppointments[0]).to.be.exist;
+    expect(state.context.questionnaire.id).to.exist;
   });
 });
